Extract tests to own files
